Develop

+++ routes/users/userknowledge.js
@@ -5,6 +5,9 @@ const modelPath = modPath.MODEL_PATH;
 const User = require(modelPath + 'usermodel')
 
 module.exports.addKnowledgeRoute = function(req, res, next) {
+  if ('' + req.user._id !== '' + req.params.userid) {
+    return res.status(401).json({success: false, msg: 'Not authorized!'})
+  }
   User.addKnowledge(req.params.userid, StringUtils.urlToName(req.params.subject), (err, savedUser) => {
     if (savedUser) {
       return res.json({success: true, msg: "Add subject to user knowledge"})
@@ -15,6 +18,9 @@ module.exports.addKnowledgeRoute = function(req, res, next) {
 }
 
 module.exports.removeKnowledgeRoute = function(req, res, next) {
+  if ('' + req.user._id !== '' + req.params.userid) {
+    return res.status(401).json({success: false, msg: 'Not authorized!'})
+  }
   User.deleteKnowledge(req.params.userid, StringUtils.urlToName(req.params.subject), (err, savedUser) => {
     if (savedUser) {
       return res.json({success: true, msg: "Deleted subject from user knowledge"})
@@ -25,6 +31,9 @@ module.exports.removeKnowledgeRoute = function(req, res, next) {
 }
 
 module.exports.updateKnowledgeRoute = function(req, res, next) {
+  if ('' + req.user._id !== '' + req.params.userid) {
+    return res.status(401).json({success: false, msg: 'Not authorized!'})
+  }
   User.updateKnowledge(req.params.userid, req.body.knowledge, (err, savedUser) => {
     if (savedUser) {
       return res.json({success: true, msg: "Successfully updated user"})
